requirements.txt
asgiref==3.3.1
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.0
django-3-jet==1.0.8
django-crispy-forms==1.10.0
django-import-export==2.5.0
django-suit==0.2.28
et-xmlfile==1.0.1
feedparser==6.0.2
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.5
pkg-resources==0.0.0
pytz==2020.5
PyYAML==5.3.1
sgmllib3k==1.0.0
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
xlrd==2.0.1
xlwt==1.3.0
